HabitReleaseFix : Fix: 테스트 과정중에 발견한 오류 수정2

+++ src/main/java/com/sollertia/habit/domain/completedhabbit/SimpleHabitVo.java
@@ -7,7 +7,7 @@ import lombok.Setter;
 @Setter
 public class SimpleHabitVo {
     private String title;
-    private int achievement;
+    private Long achievement;
 
     public SimpleHabitVo(CompletedHabit completedHabit) {
         this.title = completedHabit.getTitle();
